sub/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
scikit-learn==1.7.1
kafka-python==2.2.15
pymongo==4.14.1